requirements.txt
grpcio==1.59.0
grpcio-tools==1.59.0
PyMuPDF==1.22.5
scikit-learn==1.3.1
numpy==1.26.0
requests==2.31.0
PyJWT
img2pdf
pdf2image
opencv-python-headless
pillow
pytesseract